Rename `Executor` to `Driver` (#50)
